Added exceptions

+++ main.py
@@ -1,9 +1,15 @@
-from flask import Flask, request
+from flask import Flask, request, jsonify
 import Remote.RemoteLogic as rem
 from DataBase import WorkWithDB as wwdb
 import datetime
 
 app = Flask(__name__)
+
+
+def my403(text):
+    respons = jsonify({'error': text})
+    respons.status_code = 403
+    return respons
 
 
 @app.route('/')
@@ -14,6 +20,15 @@ def hello():
 @app.route('/connect', methods=['POST'])
 def Connect():
     request_data = request.get_json()
+
+    if 'host' not in request_data:
+        return my403('There is no host in the request')
+    if 'port' not in request_data:
+        return my403('There is no port in the request')
+    if 'username' not in request_data:
+        return my403('There is no username in the request')
+    if 'command' not in request_data:
+        return my403('There is no command in the request')
 
     use_ssh_key = request_data["UseSSHKey"] if 'UseSSHKey' in request_data else None
     hostname = request_data['host']
@@ -28,7 +43,7 @@ def Connect():
     else:
         if rem.first_connect(hostname):
             if password is None:
-                raise Exception("It is impossible to establish SSH connect via keys without password for the first time")
+                return my403('It is impossible to establish SSH connect via keys without password for the first time')
 
             rem.keygen(hostname, username, password)
 
@@ -46,9 +61,13 @@ def Connect():
 
     return result
 
+
 @app.route('/gethistory', methods=['GET'])
 def GetHistory():
     request_data = request.get_json()
+
+    if 'userGUID' not in request_data:
+        return my403('There is no userGUID in the request')
 
     userGUID = request_data['userGUID']
 
@@ -60,6 +79,17 @@ def GetHistory():
 @app.route('/savemachine', methods=['POST'])
 def SaveMachine():
     request_data = request.get_json()
+
+    if 'userGUID' not in request_data:
+        return my403('There is no userGUID in the request')
+    if 'host' not in request_data:
+        return my403('There is no host in the request')
+    if 'port' not in request_data:
+        return my403('There is no port in the request')
+    if 'username' not in request_data:
+        return my403('There is no username in the request')
+    if 'command' not in request_data:
+        return my403('There is no command in the request')
 
     userGUID = request_data['userGUID']
     host = request_data['host']
@@ -74,6 +104,13 @@ def SaveMachine():
 def SaveUser():
     request_data = request.get_json()
 
+    if 'login' not in request_data:
+        return my403('There is no login in the request')
+    if 'name' not in request_data:
+        return my403('There is no name in the request')
+    if 'password' not in request_data:
+        return my403('There is no password in the request')
+
     login = request_data['login']
     name = request_data['name']
     password = request_data['password']
@@ -84,6 +121,11 @@ def SaveUser():
 @app.route('/updateuser', methods=['POST'])
 def UpdateUser():
     request_data = request.get_json()
+
+    if 'userGUID' not in request_data:
+        return my403('There is no userGUID in the request')
+    if 'login' not in request_data and 'name' not in request_data and 'password' not in request_data:
+        return my403('The query has no parameters to change(login, name, password)')
 
     guid = request_data['userGUID']
     login = request_data['login'] if 'login' in request_data else None
@@ -97,6 +139,11 @@ def UpdateUser():
 def UpdateMachine():
     request_data = request.get_json()
 
+    if 'machineGUID' not in request_data:
+        return my403('There is no machineGUID in the request')
+    if 'host' not in request_data and 'port' not in request_data and 'password' not in request_data:
+        return my403('The query has no parameters to change(host, port, password)')
+
     guid = request_data['machineGUID']
     host = request_data['host'] if 'host' in request_data else None
     port = request_data['port'] if 'port' in request_data else None
@@ -109,6 +156,9 @@ def UpdateMachine():
 def DeleteUser():
     request_data = request.get_json()
 
+    if 'userGUID' not in request_data:
+        return my403('There is no userGUID in the request')
+
     guid = request_data['userGUID']
 
     wwdb.DeleteUser(guid)
@@ -118,9 +168,13 @@ def DeleteUser():
 def DeleteMachine():
     request_data = request.get_json()
 
+    if 'machineGUID' not in request_data:
+        return my403('There is no machineGUID in the request')
+
     guid = request_data['machineGUID']
 
     wwdb.DeleteMachine(guid)
+
 
 if __name__ == '__main__':
     app.run()
